backend/requirements.txt
# Bet_Hope Requirements
# Django + Python AI Football Prediction Platform

# =============================================================================
# CORE DJANGO
# =============================================================================
Django>=5.0,<5.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5,<25.0
django-extensions>=3.2.3
drf-spectacular>=0.27.0  # OpenAPI schema generation

# =============================================================================
# DATABASE
# =============================================================================
psycopg2-binary>=2.9.9  # PostgreSQL adapter
django-environ>=0.11.2   # Environment variable management
dj-database-url>=2.1.0   # Database URL parsing

# =============================================================================
# AUTHENTICATION
# =============================================================================
djangorestframework-simplejwt>=5.3.1
django-allauth>=0.59.0

# =============================================================================
# ASYNC TASKS
# =============================================================================
celery>=5.3.4
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
redis>=5.0.1
flower>=2.0.1  # Celery monitoring

# =============================================================================
# MACHINE LEARNING
# =============================================================================
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.2
xgboost>=2.0.2
joblib>=1.3.2
scipy>=1.11.4

# Optional: Deep Learning (uncomment if needed)
# torch>=2.1.0
# pytorch-lightning>=2.1.0

# Hyperparameter tuning
optuna>=3.4.0

# =============================================================================
# AI / NLP
# =============================================================================
sentence-transformers>=2.2.2
transformers>=4.35.0
huggingface-hub>=0.19.0
tiktoken>=0.5.2

# AI Provider SDKs
openai>=1.3.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# =============================================================================
# VECTOR DATABASE
# =============================================================================
pgvector>=0.2.4  # PostgreSQL vector extension

# Alternative: Qdrant (uncomment if using)
# qdrant-client>=1.6.0

# =============================================================================
# HTTP & API CLIENTS
# =============================================================================
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.9.0

# =============================================================================
# DATA VALIDATION
# =============================================================================
pydantic>=2.5.0
pydantic-settings>=2.1.0

# =============================================================================
# UTILITIES
# =============================================================================
python-dateutil>=2.8.2
pytz>=2023.3
python-slugify>=8.0.1
Pillow>=10.1.0  # Image processing

# =============================================================================
# CACHING
# =============================================================================
django-redis>=5.4.0

# =============================================================================
# LOGGING & MONITORING
# =============================================================================
sentry-sdk>=1.34.0
structlog>=23.2.0
python-json-logger>=2.0.7

# =============================================================================
# TESTING
# =============================================================================
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
factory-boy>=3.3.0
faker>=20.0.0
httpx>=0.25.0  # For async test client
pytest-xdist>=3.5.0  # Parallel test execution

# =============================================================================
# CODE QUALITY
# =============================================================================
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.6.0

# Type stubs
django-stubs>=4.2.0
djangorestframework-stubs>=3.14.0
types-requests>=2.31.0
pandas-stubs>=2.1.0

# =============================================================================
# PRODUCTION
# =============================================================================
gunicorn>=21.2.0
uvicorn>=0.24.0
whitenoise>=6.6.0  # Static file serving

# =============================================================================
# DOCUMENTATION
# =============================================================================
mkdocs>=1.5.0
mkdocs-material>=9.5.0
mkdocstrings>=0.24.0

# =============================================================================
# DEVELOPMENT
# =============================================================================
ipython>=8.18.0
django-debug-toolbar>=4.2.0
watchfiles>=0.21.0  # File watching for development
